move CreateValuesDictionary from CopyGuidCommand to VSPackage.
