+++ CreateGUIDVSPlugin/CopyGuidCommand.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel.Design;
 using System.Globalization;
 using System.Text;
@@ -136,33 +135,6 @@ namespace CreateGUIDVSPlugin
 #endif
 
         /// <summary>
-        /// Create a dictionary for the template values
-        /// </summary>
-        /// <returns></returns>
-        private Dictionary<string, string> CreateValuesDictionary(Guid guid)
-        {
-            var values = new Dictionary<string, string>();
-            foreach (VariableManager variableManager in Template.Variables)
-            {
-                values[variableManager.Variable] = string.Empty;
-            }
-
-            //
-            // see https://msdn.microsoft.com/library/97af8hh4(v=vs.110).aspx about the parameter of Guid.ToString().
-            // see https://msdn.microsoft.com/library/system.guid(v=vs.110).aspx
-            // see 'Reference Source' link in the above site.
-            //
-            var lowerWithHyphens = guid.ToString("D");
-            values[Template.VariableLowerCaseGuidWithHyphens] = lowerWithHyphens;
-            values[Template.VariableUpperCaseGuidWithHyphens] = lowerWithHyphens.ToUpper();
-
-            var lowerWithoutHyphens = guid.ToString("N");
-            values[Template.VariableLowerCaseGuidWithoutHyphens] = lowerWithoutHyphens;
-            values[Template.VariableUpperCaseGuidWithoutHyphens] = lowerWithoutHyphens.ToUpper();
-            return values;
-        }
-
-        /// <summary>
         /// This function is the callback used to execute the command when the menu item is clicked.
         /// See the constructor to see how the menu item is associated with this function using
         /// OleMenuCommandService service and MenuCommand class.
@@ -177,7 +149,7 @@ namespace CreateGUIDVSPlugin
             if (activeDocument != null)
             {
                 var guid = Guid.NewGuid();
-                var values = CreateValuesDictionary(guid);
+                var values = this.package.CreateValuesDictionary(guid);
                 var configuration = this.package.GetConfiguration();
                 var formatString = configuration.FormatString;
