+++ CreateGUIDVSPlugin/VSPackage.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.ComponentModel.Design;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
@@ -100,6 +101,33 @@ namespace CreateGUIDVSPlugin
         /// Property For OutputWindow
         /// </summary>
         public EnvDTE.OutputWindowPane OutputPane { get; private set; }
+
+        /// <summary>
+        /// Create a dictionary for the template values
+        /// </summary>
+        /// <returns></returns>
+        public Dictionary<string, string> CreateValuesDictionary(Guid guid)
+        {
+            var values = new Dictionary<string, string>();
+            foreach (VariableManager variableManager in Template.Variables)
+            {
+                values[variableManager.Variable] = string.Empty;
+            }
+
+            //
+            // see https://msdn.microsoft.com/library/97af8hh4(v=vs.110).aspx about the parameter of Guid.ToString().
+            // see https://msdn.microsoft.com/library/system.guid(v=vs.110).aspx
+            // see 'Reference Source' link in the above site.
+            //
+            var lowerWithHyphens = guid.ToString("D");
+            values[Template.VariableLowerCaseGuidWithHyphens] = lowerWithHyphens;
+            values[Template.VariableUpperCaseGuidWithHyphens] = lowerWithHyphens.ToUpper();
+
+            var lowerWithoutHyphens = guid.ToString("N");
+            values[Template.VariableLowerCaseGuidWithoutHyphens] = lowerWithoutHyphens;
+            values[Template.VariableUpperCaseGuidWithoutHyphens] = lowerWithoutHyphens.ToUpper();
+            return values;
+        }
 
         #region Package Members
